Adds enemy ship

+++ app/src/main/java/network/iut/org/flappydragon/interfaces/Collidable.java
@@ -6,6 +6,6 @@ package network.iut.org.flappydragon.interfaces;
 
 public interface Collidable {
 
-    public boolean collideWith(double x, double y);
+    public boolean collideWith(float x, float y);
 
 }
